get_head_size

+++ src/obfs/obfsutil.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int get_head_size(const uint8_t *plaindata, int size, int def_size);
+size_t get_head_size(const uint8_t *plaindata, size_t size, size_t def_size);
 
 void init_shift128plus(void);
